Replace initial database migration with a new re-init migration
