+++ database/versions/097d0a060aef_re_init.py
@@ -1,0 +1,26 @@
+"""re-init
+
+Revision ID: 097d0a060aef
+Revises: 
+Create Date: 2024-12-17 15:59:24.309774
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = '097d0a060aef'
+down_revision: Union[str, None] = None
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    pass
+
+
+def downgrade() -> None:
+    pass
