--- database/versions/cbc697a95795_init_database.py
@@ -1,36 +1,0 @@
-"""init database
-
-Revision ID: cbc697a95795
-Revises: 
-Create Date: 2024-07-10 11:11:05.935051
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'cbc697a95795'
-down_revision: Union[str, None] = None
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('users',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('name', sa.String(255), nullable=True),
-    sa.Column('fullname', sa.String(255), nullable=True),
-    sa.Column('nickname', sa.String(255), nullable=True),
-    sa.PrimaryKeyConstraint('id')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('users')
-    # ### end Alembic commands ###
