--- database/versions/fa9c8b7e809a_fix_type.py
@@ -1,531 +1,0 @@
-"""Fix Type
-
-Revision ID: fa9c8b7e809a
-Revises: cbc697a95795
-Create Date: 2024-07-27 13:50:31.555360
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import mysql
-
-# revision identifiers, used by Alembic.
-revision: str = 'fa9c8b7e809a'
-down_revision: Union[str, None] = 'cbc697a95795'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index('name', table_name='help_topic')
-    op.drop_table('help_topic')
-    op.drop_index('name', table_name='help_keyword')
-    op.drop_table('help_keyword')
-    op.drop_table('help_relation')
-    op.drop_table('db')
-    op.drop_table('procs_priv')
-    op.drop_table('columns_priv')
-    op.drop_table('tables_priv')
-    op.drop_table('plugin')
-    op.drop_table('innodb_table_stats')
-    op.drop_index('Host', table_name='roles_mapping')
-    op.drop_table('roles_mapping')
-    op.drop_table('gtid_slave_pos')
-    op.drop_table('table_stats')
-    op.drop_table('time_zone_transition_type')
-    op.drop_table('general_log')
-    op.drop_table('index_stats')
-    op.drop_table('slow_log')
-    op.drop_table('proxies_priv')
-    op.drop_table('global_priv')
-    op.drop_table('event')
-    op.drop_table('func')
-    op.drop_table('time_zone')
-    op.drop_table('servers')
-    op.drop_table('time_zone_name')
-    op.drop_table('column_stats')
-    op.drop_table('time_zone_leap_second')
-    op.drop_table('proc')
-    op.drop_index('commit_id', table_name='transaction_registry')
-    op.drop_table('transaction_registry')
-    op.drop_index('name', table_name='help_category')
-    op.drop_table('help_category')
-    op.drop_table('time_zone_transition')
-    op.drop_table('innodb_index_stats')
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('innodb_index_stats',
-    sa.Column('database_name', mysql.VARCHAR(length=64), nullable=False),
-    sa.Column('table_name', mysql.VARCHAR(length=199), nullable=False),
-    sa.Column('index_name', mysql.VARCHAR(length=64), nullable=False),
-    sa.Column('last_update', mysql.TIMESTAMP(), server_default=sa.text('current_timestamp() ON UPDATE current_timestamp()'), nullable=False),
-    sa.Column('stat_name', mysql.VARCHAR(length=64), nullable=False),
-    sa.Column('stat_value', mysql.BIGINT(display_width=20, unsigned=True), autoincrement=False, nullable=False),
-    sa.Column('sample_size', mysql.BIGINT(display_width=20, unsigned=True), autoincrement=False, nullable=True),
-    sa.Column('stat_description', mysql.VARCHAR(length=1024), nullable=False),
-    sa.PrimaryKeyConstraint('database_name', 'table_name', 'index_name', 'stat_name'),
-    mysql_collate='utf8mb3_bin',
-    mysql_default_charset='utf8mb3',
-    mysql_engine='InnoDB'
-    )
-    op.create_table('time_zone_transition',
-    sa.Column('Time_zone_id', mysql.INTEGER(display_width=10, unsigned=True), autoincrement=False, nullable=False),
-    sa.Column('Transition_time', mysql.BIGINT(display_width=20), autoincrement=False, nullable=False),
-    sa.Column('Transition_type_id', mysql.INTEGER(display_width=10, unsigned=True), autoincrement=False, nullable=False),
-    sa.PrimaryKeyConstraint('Time_zone_id', 'Transition_time'),
-    comment='Time zone transitions',
-    mysql_checksum='1',
-    mysql_collate='utf8mb3_general_ci',
-    mysql_comment='Time zone transitions',
-    mysql_default_charset='utf8mb3',
-    mysql_engine='Aria'
-    )
-    op.create_table('help_category',
-    sa.Column('help_category_id', mysql.SMALLINT(display_width=5, unsigned=True), autoincrement=False, nullable=False),
-    sa.Column('name', mysql.CHAR(length=64), nullable=False),
-    sa.Column('parent_category_id', mysql.SMALLINT(display_width=5, unsigned=True), autoincrement=False, nullable=True),
-    sa.Column('url', mysql.TEXT(), nullable=False),
-    sa.PrimaryKeyConstraint('help_category_id'),
-    comment='help categories',
-    mysql_checksum='1',
-    mysql_collate='utf8mb3_general_ci',
-    mysql_comment='help categories',
-    mysql_default_charset='utf8mb3',
-    mysql_engine='Aria'
-    )
-    op.create_index('name', 'help_category', ['name'], unique=True)
-    op.create_table('transaction_registry',
-    sa.Column('transaction_id', mysql.BIGINT(display_width=20, unsigned=True), autoincrement=False, nullable=False),
-    sa.Column('commit_id', mysql.BIGINT(display_width=20, unsigned=True), autoincrement=False, nullable=False),
-    sa.Column('begin_timestamp', mysql.TIMESTAMP(fsp=6), server_default=sa.text("'0000-00-00 00:00:00.000000'"), nullable=False),
-    sa.Column('commit_timestamp', mysql.TIMESTAMP(fsp=6), server_default=sa.text("'0000-00-00 00:00:00.000000'"), nullable=False),
-    sa.Column('isolation_level', mysql.ENUM('READ-UNCOMMITTED', 'READ-COMMITTED', 'REPEATABLE-READ', 'SERIALIZABLE'), nullable=False),
-    sa.PrimaryKeyConstraint('transaction_id'),
-    mysql_collate='utf8mb3_bin',
-    mysql_default_charset='utf8mb3',
-    mysql_engine='InnoDB'
-    )
-    op.create_index('commit_id', 'transaction_registry', ['commit_id'], unique=True)
-    op.create_table('proc',
-    sa.Column('db', mysql.CHAR(charset='utf8mb3', collation='utf8mb3_bin', length=64), server_default=sa.text("''"), nullable=False),
-    sa.Column('name', mysql.CHAR(length=64), server_default=sa.text("''"), nullable=False),
-    sa.Column('type', mysql.ENUM('FUNCTION', 'PROCEDURE', 'PACKAGE', 'PACKAGE BODY'), nullable=False),
-    sa.Column('specific_name', mysql.CHAR(length=64), server_default=sa.text("''"), nullable=False),
-    sa.Column('language', mysql.ENUM('SQL'), server_default=sa.text("'SQL'"), nullable=False),
-    sa.Column('sql_data_access', mysql.ENUM('CONTAINS_SQL', 'NO_SQL', 'READS_SQL_DATA', 'MODIFIES_SQL_DATA'), server_default=sa.text("'CONTAINS_SQL'"), nullable=False),
-    sa.Column('is_deterministic', mysql.ENUM('YES', 'NO'), server_default=sa.text("'NO'"), nullable=False),
-    sa.Column('security_type', mysql.ENUM('INVOKER', 'DEFINER'), server_default=sa.text("'DEFINER'"), nullable=False),
-    sa.Column('param_list', sa.BLOB(), nullable=False),
-    sa.Column('returns', mysql.LONGBLOB(), nullable=False),
-    sa.Column('body', mysql.LONGBLOB(), nullable=False),
-    sa.Column('definer', mysql.VARCHAR(charset='utf8mb3', collation='utf8mb3_bin', length=384), server_default=sa.text("''"), nullable=False),
-    sa.Column('created', mysql.TIMESTAMP(), server_default=sa.text('current_timestamp() ON UPDATE current_timestamp()'), nullable=False),
-    sa.Column('modified', mysql.TIMESTAMP(), server_default=sa.text("'0000-00-00 00:00:00'"), nullable=False),
-    sa.Column('sql_mode', mysql.SET('REAL_AS_FLOAT', 'PIPES_AS_CONCAT', 'ANSI_QUOTES', 'IGNORE_SPACE', 'IGNORE_BAD_TABLE_OPTIONS', 'ONLY_FULL_GROUP_BY', 'NO_UNSIGNED_SUBTRACTION', 'NO_DIR_IN_CREATE', 'POSTGRESQL', 'ORACLE', 'MSSQL', 'DB2', 'MAXDB', 'NO_KEY_OPTIONS', 'NO_TABLE_OPTIONS', 'NO_FIELD_OPTIONS', 'MYSQL323', 'MYSQL40', 'ANSI', 'NO_AUTO_VALUE_ON_ZERO', 'NO_BACKSLASH_ESCAPES', 'STRICT_TRANS_TABLES', 'STRICT_ALL_TABLES', 'NO_ZERO_IN_DATE', 'NO_ZERO_DATE', 'INVALID_DATES', 'ERROR_FOR_DIVISION_BY_ZERO', 'TRADITIONAL', 'NO_AUTO_CREATE_USER', 'HIGH_NOT_PRECEDENCE', 'NO_ENGINE_SUBSTITUTION', 'PAD_CHAR_TO_FULL_LENGTH', 'EMPTY_STRING_IS_NULL', 'SIMULTANEOUS_ASSIGNMENT', 'TIME_ROUND_FRACTIONAL'), server_default=sa.text("''"), nullable=False),
-    sa.Column('comment', mysql.TEXT(charset='utf8mb3', collation='utf8mb3_bin'), nullable=False),
-    sa.Column('character_set_client', mysql.CHAR(charset='utf8mb3', collation='utf8mb3_bin', length=32), nullable=True),
-    sa.Column('collation_connection', mysql.CHAR(charset='utf8mb3', collation='utf8mb3_bin', length=64), nullable=True),
-    sa.Column('db_collation', mysql.CHAR(charset='utf8mb3', collation='utf8mb3_bin', length=64), nullable=True),
-    sa.Column('body_utf8', mysql.LONGBLOB(), nullable=True),
-    sa.Column('aggregate', mysql.ENUM('NONE', 'GROUP'), server_default=sa.text("'NONE'"), nullable=False),
-    sa.PrimaryKeyConstraint('db', 'name', 'type'),
-    comment='Stored Procedures',
-    mysql_checksum='1',
-    mysql_collate='utf8mb3_general_ci',
-    mysql_comment='Stored Procedures',
-    mysql_default_charset='utf8mb3',
-    mysql_engine='Aria'
-    )
-    op.create_table('time_zone_leap_second',
-    sa.Column('Transition_time', mysql.BIGINT(display_width=20), autoincrement=False, nullable=False),
-    sa.Column('Correction', mysql.INTEGER(display_width=11), autoincrement=False, nullable=False),
-    sa.PrimaryKeyConstraint('Transition_time'),
-    comment='Leap seconds information for time zones',
-    mysql_checksum='1',
-    mysql_collate='utf8mb3_general_ci',
-    mysql_comment='Leap seconds information for time zones',
-    mysql_default_charset='utf8mb3',
-    mysql_engine='Aria'
-    )
-    op.create_table('column_stats',
-    sa.Column('db_name', mysql.VARCHAR(length=64), nullable=False),
-    sa.Column('table_name', mysql.VARCHAR(length=64), nullable=False),
-    sa.Column('column_name', mysql.VARCHAR(length=64), nullable=False),
-    sa.Column('min_value', sa.VARBINARY(length=255), nullable=True),
-    sa.Column('max_value', sa.VARBINARY(length=255), nullable=True),
-    sa.Column('nulls_ratio', mysql.DECIMAL(precision=12, scale=4), nullable=True),
-    sa.Column('avg_length', mysql.DECIMAL(precision=12, scale=4), nullable=True),
-    sa.Column('avg_frequency', mysql.DECIMAL(precision=12, scale=4), nullable=True),
-    sa.Column('hist_size', mysql.TINYINT(display_width=3, unsigned=True), autoincrement=False, nullable=True),
-    sa.Column('hist_type', mysql.ENUM('SINGLE_PREC_HB', 'DOUBLE_PREC_HB', 'JSON_HB'), nullable=True),
-    sa.Column('histogram', mysql.LONGBLOB(), nullable=True),
-    sa.PrimaryKeyConstraint('db_name', 'table_name', 'column_name'),
-    comment='Statistics on Columns',
-    mysql_checksum='1',
-    mysql_collate='utf8mb3_bin',
-    mysql_comment='Statistics on Columns',
-    mysql_default_charset='utf8mb3',
-    mysql_engine='Aria'
-    )
-    op.create_table('time_zone_name',
-    sa.Column('Name', mysql.CHAR(length=64), nullable=False),
-    sa.Column('Time_zone_id', mysql.INTEGER(display_width=10, unsigned=True), autoincrement=False, nullable=False),
-    sa.PrimaryKeyConstraint('Name'),
-    comment='Time zone names',
-    mysql_checksum='1',
-    mysql_collate='utf8mb3_general_ci',
-    mysql_comment='Time zone names',
-    mysql_default_charset='utf8mb3',
-    mysql_engine='Aria'
-    )
-    op.create_table('servers',
-    sa.Column('Server_name', mysql.CHAR(length=64), server_default=sa.text("''"), nullable=False),
-    sa.Column('Host', mysql.VARCHAR(length=2048), server_default=sa.text("''"), nullable=False),
-    sa.Column('Db', mysql.CHAR(length=64), server_default=sa.text("''"), nullable=False),
-    sa.Column('Username', mysql.CHAR(length=128), server_default=sa.text("''"), nullable=False),
-    sa.Column('Password', mysql.CHAR(length=64), server_default=sa.text("''"), nullable=False),
-    sa.Column('Port', mysql.INTEGER(display_width=4), server_default=sa.text('0'), autoincrement=False, nullable=False),
-    sa.Column('Socket', mysql.CHAR(length=64), server_default=sa.text("''"), nullable=False),
-    sa.Column('Wrapper', mysql.CHAR(length=64), server_default=sa.text("''"), nullable=False),
-    sa.Column('Owner', mysql.VARCHAR(length=512), server_default=sa.text("''"), nullable=False),
-    sa.PrimaryKeyConstraint('Server_name'),
-    comment='MySQL Foreign Servers table',
-    mysql_checksum='1',
-    mysql_collate='utf8mb3_general_ci',
-    mysql_comment='MySQL Foreign Servers table',
-    mysql_default_charset='utf8mb3',
-    mysql_engine='Aria'
-    )
-    op.create_table('time_zone',
-    sa.Column('Time_zone_id', mysql.INTEGER(display_width=10, unsigned=True), autoincrement=True, nullable=False),
-    sa.Column('Use_leap_seconds', mysql.ENUM('Y', 'N'), server_default=sa.text("'N'"), nullable=False),
-    sa.PrimaryKeyConstraint('Time_zone_id'),
-    comment='Time zones',
-    mysql_checksum='1',
-    mysql_collate='utf8mb3_general_ci',
-    mysql_comment='Time zones',
-    mysql_default_charset='utf8mb3',
-    mysql_engine='Aria'
-    )
-    op.create_table('func',
-    sa.Column('name', mysql.CHAR(length=64), server_default=sa.text("''"), nullable=False),
-    sa.Column('ret', mysql.TINYINT(display_width=1), server_default=sa.text('0'), autoincrement=False, nullable=False),
-    sa.Column('dl', mysql.CHAR(length=128), server_default=sa.text("''"), nullable=False),
-    sa.Column('type', mysql.ENUM('function', 'aggregate'), nullable=False),
-    sa.PrimaryKeyConstraint('name'),
-    comment='User defined functions',
-    mysql_checksum='1',
-    mysql_collate='utf8mb3_bin',
-    mysql_comment='User defined functions',
-    mysql_default_charset='utf8mb3',
-    mysql_engine='Aria'
-    )
-    op.create_table('event',
-    sa.Column('db', mysql.CHAR(charset='utf8mb3', collation='utf8mb3_bin', length=64), server_default=sa.text("''"), nullable=False),
-    sa.Column('name', mysql.CHAR(length=64), server_default=sa.text("''"), nullable=False),
-    sa.Column('body', mysql.LONGBLOB(), nullable=False),
-    sa.Column('definer', mysql.VARCHAR(charset='utf8mb3', collation='utf8mb3_bin', length=384), server_default=sa.text("''"), nullable=False),
-    sa.Column('execute_at', mysql.DATETIME(), nullable=True),
-    sa.Column('interval_value', mysql.INTEGER(display_width=11), autoincrement=False, nullable=True),
-    sa.Column('interval_field', mysql.ENUM('YEAR', 'QUARTER', 'MONTH', 'DAY', 'HOUR', 'MINUTE', 'WEEK', 'SECOND', 'MICROSECOND', 'YEAR_MONTH', 'DAY_HOUR', 'DAY_MINUTE', 'DAY_SECOND', 'HOUR_MINUTE', 'HOUR_SECOND', 'MINUTE_SECOND', 'DAY_MICROSECOND', 'HOUR_MICROSECOND', 'MINUTE_MICROSECOND', 'SECOND_MICROSECOND'), nullable=True),
-    sa.Column('created', mysql.TIMESTAMP(), server_default=sa.text('current_timestamp() ON UPDATE current_timestamp()'), nullable=False),
-    sa.Column('modified', mysql.TIMESTAMP(), server_default=sa.text("'0000-00-00 00:00:00'"), nullable=False),
-    sa.Column('last_executed', mysql.DATETIME(), nullable=True),
-    sa.Column('starts', mysql.DATETIME(), nullable=True),
-    sa.Column('ends', mysql.DATETIME(), nullable=True),
-    sa.Column('status', mysql.ENUM('ENABLED', 'DISABLED', 'SLAVESIDE_DISABLED'), server_default=sa.text("'ENABLED'"), nullable=False),
-    sa.Column('on_completion', mysql.ENUM('DROP', 'PRESERVE'), server_default=sa.text("'DROP'"), nullable=False),
-    sa.Column('sql_mode', mysql.SET('REAL_AS_FLOAT', 'PIPES_AS_CONCAT', 'ANSI_QUOTES', 'IGNORE_SPACE', 'IGNORE_BAD_TABLE_OPTIONS', 'ONLY_FULL_GROUP_BY', 'NO_UNSIGNED_SUBTRACTION', 'NO_DIR_IN_CREATE', 'POSTGRESQL', 'ORACLE', 'MSSQL', 'DB2', 'MAXDB', 'NO_KEY_OPTIONS', 'NO_TABLE_OPTIONS', 'NO_FIELD_OPTIONS', 'MYSQL323', 'MYSQL40', 'ANSI', 'NO_AUTO_VALUE_ON_ZERO', 'NO_BACKSLASH_ESCAPES', 'STRICT_TRANS_TABLES', 'STRICT_ALL_TABLES', 'NO_ZERO_IN_DATE', 'NO_ZERO_DATE', 'INVALID_DATES', 'ERROR_FOR_DIVISION_BY_ZERO', 'TRADITIONAL', 'NO_AUTO_CREATE_USER', 'HIGH_NOT_PRECEDENCE', 'NO_ENGINE_SUBSTITUTION', 'PAD_CHAR_TO_FULL_LENGTH', 'EMPTY_STRING_IS_NULL', 'SIMULTANEOUS_ASSIGNMENT', 'TIME_ROUND_FRACTIONAL'), server_default=sa.text("''"), nullable=False),
-    sa.Column('comment', mysql.CHAR(charset='utf8mb3', collation='utf8mb3_bin', length=64), server_default=sa.text("''"), nullable=False),
-    sa.Column('originator', mysql.INTEGER(display_width=10, unsigned=True), autoincrement=False, nullable=False),
-    sa.Column('time_zone', mysql.CHAR(charset='latin1', collation='latin1_swedish_ci', length=64), server_default=sa.text("'SYSTEM'"), nullable=False),
-    sa.Column('character_set_client', mysql.CHAR(charset='utf8mb3', collation='utf8mb3_bin', length=32), nullable=True),
-    sa.Column('collation_connection', mysql.CHAR(charset='utf8mb3', collation='utf8mb3_bin', length=64), nullable=True),
-    sa.Column('db_collation', mysql.CHAR(charset='utf8mb3', collation='utf8mb3_bin', length=64), nullable=True),
-    sa.Column('body_utf8', mysql.LONGBLOB(), nullable=True),
-    sa.PrimaryKeyConstraint('db', 'name'),
-    comment='Events',
-    mysql_checksum='1',
-    mysql_collate='utf8mb3_general_ci',
-    mysql_comment='Events',
-    mysql_default_charset='utf8mb3',
-    mysql_engine='Aria'
-    )
-    op.create_table('global_priv',
-    sa.Column('Host', mysql.CHAR(length=255), server_default=sa.text("''"), nullable=False),
-    sa.Column('User', mysql.CHAR(length=128), server_default=sa.text("''"), nullable=False),
-    sa.Column('Priv', mysql.LONGTEXT(charset='utf8mb4', collation='utf8mb4_bin'), server_default=sa.text("'{}'"), nullable=False),
-    sa.PrimaryKeyConstraint('Host', 'User'),
-    comment='Users and global privileges',
-    mysql_checksum='1',
-    mysql_collate='utf8mb3_bin',
-    mysql_comment='Users and global privileges',
-    mysql_default_charset='utf8mb3',
-    mysql_engine='Aria'
-    )
-    op.create_table('proxies_priv',
-    sa.Column('Host', mysql.CHAR(length=255), server_default=sa.text("''"), nullable=False),
-    sa.Column('User', mysql.CHAR(length=128), server_default=sa.text("''"), nullable=False),
-    sa.Column('Proxied_host', mysql.CHAR(length=255), server_default=sa.text("''"), nullable=False),
-    sa.Column('Proxied_user', mysql.CHAR(length=128), server_default=sa.text("''"), nullable=False),
-    sa.Column('With_grant', mysql.TINYINT(display_width=1), server_default=sa.text('0'), autoincrement=False, nullable=False),
-    sa.Column('Grantor', mysql.VARCHAR(length=384), server_default=sa.text("''"), nullable=False),
-    sa.Column('Timestamp', mysql.TIMESTAMP(), server_default=sa.text('current_timestamp() ON UPDATE current_timestamp()'), nullable=False),
-    sa.PrimaryKeyConstraint('Host', 'User', 'Proxied_host', 'Proxied_user'),
-    comment='User proxy privileges',
-    mysql_checksum='1',
-    mysql_collate='utf8mb3_bin',
-    mysql_comment='User proxy privileges',
-    mysql_default_charset='utf8mb3',
-    mysql_engine='Aria'
-    )
-    op.create_table('slow_log',
-    sa.Column('start_time', mysql.TIMESTAMP(fsp=6), server_default=sa.text('current_timestamp(6) ON UPDATE current_timestamp(6)'), nullable=False),
-    sa.Column('user_host', mysql.MEDIUMTEXT(), nullable=False),
-    sa.Column('query_time', mysql.TIME(fsp=6), nullable=False),
-    sa.Column('lock_time', mysql.TIME(fsp=6), nullable=False),
-    sa.Column('rows_sent', mysql.BIGINT(display_width=20, unsigned=True), autoincrement=False, nullable=False),
-    sa.Column('rows_examined', mysql.BIGINT(display_width=20, unsigned=True), autoincrement=False, nullable=False),
-    sa.Column('db', mysql.VARCHAR(length=512), nullable=False),
-    sa.Column('last_insert_id', mysql.INTEGER(display_width=11), autoincrement=False, nullable=False),
-    sa.Column('insert_id', mysql.INTEGER(display_width=11), autoincrement=False, nullable=False),
-    sa.Column('server_id', mysql.INTEGER(display_width=10, unsigned=True), autoincrement=False, nullable=False),
-    sa.Column('sql_text', mysql.MEDIUMTEXT(), nullable=False),
-    sa.Column('thread_id', mysql.BIGINT(display_width=21, unsigned=True), autoincrement=False, nullable=False),
-    sa.Column('rows_affected', mysql.BIGINT(display_width=20, unsigned=True), autoincrement=False, nullable=False),
-    comment='Slow log',
-    mysql_collate='utf8mb3_general_ci',
-    mysql_comment='Slow log',
-    mysql_default_charset='utf8mb3',
-    mysql_engine='CSV'
-    )
-    op.create_table('index_stats',
-    sa.Column('db_name', mysql.VARCHAR(length=64), nullable=False),
-    sa.Column('table_name', mysql.VARCHAR(length=64), nullable=False),
-    sa.Column('index_name', mysql.VARCHAR(length=64), nullable=False),
-    sa.Column('prefix_arity', mysql.INTEGER(display_width=11, unsigned=True), autoincrement=False, nullable=False),
-    sa.Column('avg_frequency', mysql.DECIMAL(precision=12, scale=4), nullable=True),
-    sa.PrimaryKeyConstraint('db_name', 'table_name', 'index_name', 'prefix_arity'),
-    comment='Statistics on Indexes',
-    mysql_checksum='1',
-    mysql_collate='utf8mb3_bin',
-    mysql_comment='Statistics on Indexes',
-    mysql_default_charset='utf8mb3',
-    mysql_engine='Aria'
-    )
-    op.create_table('general_log',
-    sa.Column('event_time', mysql.TIMESTAMP(fsp=6), server_default=sa.text('current_timestamp(6) ON UPDATE current_timestamp(6)'), nullable=False),
-    sa.Column('user_host', mysql.MEDIUMTEXT(), nullable=False),
-    sa.Column('thread_id', mysql.BIGINT(display_width=21, unsigned=True), autoincrement=False, nullable=False),
-    sa.Column('server_id', mysql.INTEGER(display_width=10, unsigned=True), autoincrement=False, nullable=False),
-    sa.Column('command_type', mysql.VARCHAR(length=64), nullable=False),
-    sa.Column('argument', mysql.MEDIUMTEXT(), nullable=False),
-    comment='General log',
-    mysql_collate='utf8mb3_general_ci',
-    mysql_comment='General log',
-    mysql_default_charset='utf8mb3',
-    mysql_engine='CSV'
-    )
-    op.create_table('time_zone_transition_type',
-    sa.Column('Time_zone_id', mysql.INTEGER(display_width=10, unsigned=True), autoincrement=False, nullable=False),
-    sa.Column('Transition_type_id', mysql.INTEGER(display_width=10, unsigned=True), autoincrement=False, nullable=False),
-    sa.Column('Offset', mysql.INTEGER(display_width=11), server_default=sa.text('0'), autoincrement=False, nullable=False),
-    sa.Column('Is_DST', mysql.TINYINT(display_width=3, unsigned=True), server_default=sa.text('0'), autoincrement=False, nullable=False),
-    sa.Column('Abbreviation', mysql.CHAR(length=8), server_default=sa.text("''"), nullable=False),
-    sa.PrimaryKeyConstraint('Time_zone_id', 'Transition_type_id'),
-    comment='Time zone transition types',
-    mysql_checksum='1',
-    mysql_collate='utf8mb3_general_ci',
-    mysql_comment='Time zone transition types',
-    mysql_default_charset='utf8mb3',
-    mysql_engine='Aria'
-    )
-    op.create_table('table_stats',
-    sa.Column('db_name', mysql.VARCHAR(length=64), nullable=False),
-    sa.Column('table_name', mysql.VARCHAR(length=64), nullable=False),
-    sa.Column('cardinality', mysql.BIGINT(display_width=21, unsigned=True), autoincrement=False, nullable=True),
-    sa.PrimaryKeyConstraint('db_name', 'table_name'),
-    comment='Statistics on Tables',
-    mysql_checksum='1',
-    mysql_collate='utf8mb3_bin',
-    mysql_comment='Statistics on Tables',
-    mysql_default_charset='utf8mb3',
-    mysql_engine='Aria'
-    )
-    op.create_table('gtid_slave_pos',
-    sa.Column('domain_id', mysql.INTEGER(display_width=10, unsigned=True), autoincrement=False, nullable=False),
-    sa.Column('sub_id', mysql.BIGINT(display_width=20, unsigned=True), autoincrement=False, nullable=False),
-    sa.Column('server_id', mysql.INTEGER(display_width=10, unsigned=True), autoincrement=False, nullable=False),
-    sa.Column('seq_no', mysql.BIGINT(display_width=20, unsigned=True), autoincrement=False, nullable=False),
-    sa.PrimaryKeyConstraint('domain_id', 'sub_id'),
-    comment='Replication slave GTID position',
-    mysql_collate='latin1_swedish_ci',
-    mysql_comment='Replication slave GTID position',
-    mysql_default_charset='latin1',
-    mysql_engine='InnoDB'
-    )
-    op.create_table('roles_mapping',
-    sa.Column('Host', mysql.CHAR(length=255), server_default=sa.text("''"), nullable=False),
-    sa.Column('User', mysql.CHAR(length=128), server_default=sa.text("''"), nullable=False),
-    sa.Column('Role', mysql.CHAR(length=128), server_default=sa.text("''"), nullable=False),
-    sa.Column('Admin_option', mysql.ENUM('N', 'Y'), server_default=sa.text("'N'"), nullable=False),
-    comment='Granted roles',
-    mysql_checksum='1',
-    mysql_collate='utf8mb3_bin',
-    mysql_comment='Granted roles',
-    mysql_default_charset='utf8mb3',
-    mysql_engine='Aria'
-    )
-    op.create_index('Host', 'roles_mapping', ['Host', 'User', 'Role'], unique=True)
-    op.create_table('innodb_table_stats',
-    sa.Column('database_name', mysql.VARCHAR(length=64), nullable=False),
-    sa.Column('table_name', mysql.VARCHAR(length=199), nullable=False),
-    sa.Column('last_update', mysql.TIMESTAMP(), server_default=sa.text('current_timestamp() ON UPDATE current_timestamp()'), nullable=False),
-    sa.Column('n_rows', mysql.BIGINT(display_width=20, unsigned=True), autoincrement=False, nullable=False),
-    sa.Column('clustered_index_size', mysql.BIGINT(display_width=20, unsigned=True), autoincrement=False, nullable=False),
-    sa.Column('sum_of_other_index_sizes', mysql.BIGINT(display_width=20, unsigned=True), autoincrement=False, nullable=False),
-    sa.PrimaryKeyConstraint('database_name', 'table_name'),
-    mysql_collate='utf8mb3_bin',
-    mysql_default_charset='utf8mb3',
-    mysql_engine='InnoDB'
-    )
-    op.create_table('plugin',
-    sa.Column('name', mysql.VARCHAR(length=64), server_default=sa.text("''"), nullable=False),
-    sa.Column('dl', mysql.VARCHAR(length=128), server_default=sa.text("''"), nullable=False),
-    sa.PrimaryKeyConstraint('name'),
-    comment='MySQL plugins',
-    mysql_checksum='1',
-    mysql_collate='utf8mb3_general_ci',
-    mysql_comment='MySQL plugins',
-    mysql_default_charset='utf8mb3',
-    mysql_engine='Aria'
-    )
-    op.create_table('tables_priv',
-    sa.Column('Host', mysql.CHAR(length=255), server_default=sa.text("''"), nullable=False),
-    sa.Column('Db', mysql.CHAR(length=64), server_default=sa.text("''"), nullable=False),
-    sa.Column('User', mysql.CHAR(length=128), server_default=sa.text("''"), nullable=False),
-    sa.Column('Table_name', mysql.CHAR(length=64), server_default=sa.text("''"), nullable=False),
-    sa.Column('Grantor', mysql.VARCHAR(length=384), server_default=sa.text("''"), nullable=False),
-    sa.Column('Timestamp', mysql.TIMESTAMP(), server_default=sa.text('current_timestamp() ON UPDATE current_timestamp()'), nullable=False),
-    sa.Column('Table_priv', mysql.SET('Select', 'Insert', 'Update', 'Delete', 'Create', 'Drop', 'Grant', 'References', 'Index', 'Alter', 'Create View', 'Show view', 'Trigger', 'Delete versioning rows', charset='utf8mb3', collation='utf8mb3_general_ci'), server_default=sa.text("''"), nullable=False),
-    sa.Column('Column_priv', mysql.SET('Select', 'Insert', 'Update', 'References', charset='utf8mb3', collation='utf8mb3_general_ci'), server_default=sa.text("''"), nullable=False),
-    sa.PrimaryKeyConstraint('Host', 'Db', 'User', 'Table_name'),
-    comment='Table privileges',
-    mysql_checksum='1',
-    mysql_collate='utf8mb3_bin',
-    mysql_comment='Table privileges',
-    mysql_default_charset='utf8mb3',
-    mysql_engine='Aria'
-    )
-    op.create_table('columns_priv',
-    sa.Column('Host', mysql.CHAR(length=255), server_default=sa.text("''"), nullable=False),
-    sa.Column('Db', mysql.CHAR(length=64), server_default=sa.text("''"), nullable=False),
-    sa.Column('User', mysql.CHAR(length=128), server_default=sa.text("''"), nullable=False),
-    sa.Column('Table_name', mysql.CHAR(length=64), server_default=sa.text("''"), nullable=False),
-    sa.Column('Column_name', mysql.CHAR(length=64), server_default=sa.text("''"), nullable=False),
-    sa.Column('Timestamp', mysql.TIMESTAMP(), server_default=sa.text('current_timestamp() ON UPDATE current_timestamp()'), nullable=False),
-    sa.Column('Column_priv', mysql.SET('Select', 'Insert', 'Update', 'References', charset='utf8mb3', collation='utf8mb3_general_ci'), server_default=sa.text("''"), nullable=False),
-    sa.PrimaryKeyConstraint('Host', 'Db', 'User', 'Table_name', 'Column_name'),
-    comment='Column privileges',
-    mysql_checksum='1',
-    mysql_collate='utf8mb3_bin',
-    mysql_comment='Column privileges',
-    mysql_default_charset='utf8mb3',
-    mysql_engine='Aria'
-    )
-    op.create_table('procs_priv',
-    sa.Column('Host', mysql.CHAR(length=255), server_default=sa.text("''"), nullable=False),
-    sa.Column('Db', mysql.CHAR(length=64), server_default=sa.text("''"), nullable=False),
-    sa.Column('User', mysql.CHAR(length=128), server_default=sa.text("''"), nullable=False),
-    sa.Column('Routine_name', mysql.CHAR(charset='utf8mb3', collation='utf8mb3_general_ci', length=64), server_default=sa.text("''"), nullable=False),
-    sa.Column('Routine_type', mysql.ENUM('FUNCTION', 'PROCEDURE', 'PACKAGE', 'PACKAGE BODY'), nullable=False),
-    sa.Column('Grantor', mysql.VARCHAR(length=384), server_default=sa.text("''"), nullable=False),
-    sa.Column('Proc_priv', mysql.SET('Execute', 'Alter Routine', 'Grant', 'Show Create Routine', charset='utf8mb3', collation='utf8mb3_general_ci'), server_default=sa.text("''"), nullable=False),
-    sa.Column('Timestamp', mysql.TIMESTAMP(), server_default=sa.text('current_timestamp() ON UPDATE current_timestamp()'), nullable=False),
-    sa.PrimaryKeyConstraint('Host', 'Db', 'User', 'Routine_name', 'Routine_type'),
-    comment='Procedure privileges',
-    mysql_checksum='1',
-    mysql_collate='utf8mb3_bin',
-    mysql_comment='Procedure privileges',
-    mysql_default_charset='utf8mb3',
-    mysql_engine='Aria'
-    )
-    op.create_table('db',
-    sa.Column('Host', mysql.CHAR(length=255), server_default=sa.text("''"), nullable=False),
-    sa.Column('Db', mysql.CHAR(length=64), server_default=sa.text("''"), nullable=False),
-    sa.Column('User', mysql.CHAR(length=128), server_default=sa.text("''"), nullable=False),
-    sa.Column('Select_priv', mysql.ENUM('N', 'Y'), server_default=sa.text("'N'"), nullable=False),
-    sa.Column('Insert_priv', mysql.ENUM('N', 'Y'), server_default=sa.text("'N'"), nullable=False),
-    sa.Column('Update_priv', mysql.ENUM('N', 'Y'), server_default=sa.text("'N'"), nullable=False),
-    sa.Column('Delete_priv', mysql.ENUM('N', 'Y'), server_default=sa.text("'N'"), nullable=False),
-    sa.Column('Create_priv', mysql.ENUM('N', 'Y'), server_default=sa.text("'N'"), nullable=False),
-    sa.Column('Drop_priv', mysql.ENUM('N', 'Y'), server_default=sa.text("'N'"), nullable=False),
-    sa.Column('Grant_priv', mysql.ENUM('N', 'Y'), server_default=sa.text("'N'"), nullable=False),
-    sa.Column('References_priv', mysql.ENUM('N', 'Y'), server_default=sa.text("'N'"), nullable=False),
-    sa.Column('Index_priv', mysql.ENUM('N', 'Y'), server_default=sa.text("'N'"), nullable=False),
-    sa.Column('Alter_priv', mysql.ENUM('N', 'Y'), server_default=sa.text("'N'"), nullable=False),
-    sa.Column('Create_tmp_table_priv', mysql.ENUM('N', 'Y'), server_default=sa.text("'N'"), nullable=False),
-    sa.Column('Lock_tables_priv', mysql.ENUM('N', 'Y'), server_default=sa.text("'N'"), nullable=False),
-    sa.Column('Create_view_priv', mysql.ENUM('N', 'Y'), server_default=sa.text("'N'"), nullable=False),
-    sa.Column('Show_view_priv', mysql.ENUM('N', 'Y'), server_default=sa.text("'N'"), nullable=False),
-    sa.Column('Create_routine_priv', mysql.ENUM('N', 'Y'), server_default=sa.text("'N'"), nullable=False),
-    sa.Column('Alter_routine_priv', mysql.ENUM('N', 'Y'), server_default=sa.text("'N'"), nullable=False),
-    sa.Column('Execute_priv', mysql.ENUM('N', 'Y'), server_default=sa.text("'N'"), nullable=False),
-    sa.Column('Event_priv', mysql.ENUM('N', 'Y'), server_default=sa.text("'N'"), nullable=False),
-    sa.Column('Trigger_priv', mysql.ENUM('N', 'Y'), server_default=sa.text("'N'"), nullable=False),
-    sa.Column('Delete_history_priv', mysql.ENUM('N', 'Y'), server_default=sa.text("'N'"), nullable=False),
-    sa.Column('Show_create_routine_priv', mysql.ENUM('N', 'Y'), server_default=sa.text("'N'"), nullable=False),
-    sa.PrimaryKeyConstraint('Host', 'Db', 'User'),
-    comment='Database privileges',
-    mysql_checksum='1',
-    mysql_collate='utf8mb3_bin',
-    mysql_comment='Database privileges',
-    mysql_default_charset='utf8mb3',
-    mysql_engine='Aria'
-    )
-    op.create_table('help_relation',
-    sa.Column('help_topic_id', mysql.INTEGER(display_width=10, unsigned=True), autoincrement=False, nullable=False),
-    sa.Column('help_keyword_id', mysql.INTEGER(display_width=10, unsigned=True), autoincrement=False, nullable=False),
-    sa.PrimaryKeyConstraint('help_keyword_id', 'help_topic_id'),
-    comment='keyword-topic relation',
-    mysql_checksum='1',
-    mysql_collate='utf8mb3_general_ci',
-    mysql_comment='keyword-topic relation',
-    mysql_default_charset='utf8mb3',
-    mysql_engine='Aria'
-    )
-    op.create_table('help_keyword',
-    sa.Column('help_keyword_id', mysql.INTEGER(display_width=10, unsigned=True), autoincrement=False, nullable=False),
-    sa.Column('name', mysql.CHAR(length=64), nullable=False),
-    sa.PrimaryKeyConstraint('help_keyword_id'),
-    comment='help keywords',
-    mysql_checksum='1',
-    mysql_collate='utf8mb3_general_ci',
-    mysql_comment='help keywords',
-    mysql_default_charset='utf8mb3',
-    mysql_engine='Aria'
-    )
-    op.create_index('name', 'help_keyword', ['name'], unique=True)
-    op.create_table('help_topic',
-    sa.Column('help_topic_id', mysql.INTEGER(display_width=10, unsigned=True), autoincrement=False, nullable=False),
-    sa.Column('name', mysql.CHAR(length=64), nullable=False),
-    sa.Column('help_category_id', mysql.SMALLINT(display_width=5, unsigned=True), autoincrement=False, nullable=False),
-    sa.Column('description', mysql.TEXT(), nullable=False),
-    sa.Column('example', mysql.TEXT(), nullable=False),
-    sa.Column('url', mysql.TEXT(), nullable=False),
-    sa.PrimaryKeyConstraint('help_topic_id'),
-    comment='help topics',
-    mysql_checksum='1',
-    mysql_collate='utf8mb3_general_ci',
-    mysql_comment='help topics',
-    mysql_default_charset='utf8mb3',
-    mysql_engine='Aria'
-    )
-    op.create_index('name', 'help_topic', ['name'], unique=True)
-    # ### end Alembic commands ###
